Разделяет код на модули. Импортирует и экспортирует элементы.
